# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kunaverso API Documentation",
        "description": "Kunaverso Api",
        "contact": {
            "email": "oniel.gzlez@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Kunaverso API Server"
        }
    ],
    "paths": {
...
        "/user/profile-photo": {
            "post": {
                "tags": [
                    "Kunaverso"
                ],
                "summary": "Update user's photo.",
                "description": "Update user's photo.",
                "operationId": "update",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/UserPhotoRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User photo updated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PhotoUpdatedResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "array": {
                                        "summary": "An result object.",
                                        "value": {
                                            "message": "message"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "array": {
                                        "summary": "An result object.",
                                        "value": {
                                            "message": "message"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "User photo errors.",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "array": {
                                        "summary": "An result object.",
                                        "value": {
                                            "message": "Error message",
                                            "errors": {
                                                "photo": [
                                                    "Error message"
                                                ]
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/user/remove-photo": {
            "post": {
                "tags": [
                    "Kunaverso"
                ],
                "summary": "Delete user's photo.",
                "description": "Delete user's photo.",
                "operationId": "delete",
                "responses": {
                    "200": {
                        "description": "User photo deleted.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PhotoDeletedResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "array": {
                                        "summary": "An result object.",
                                        "value": {
                                            "message": "message"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "array": {
                                        "summary": "An result object.",
                                        "value": {
                                            "message": "message"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "UserPhotoRequest": {
                "title": "User Photo request",
                "description": "User Photo request body data",
                "required": [
                    "photo"
                ],
                "properties": {
                    "photo": {
                        "description": "image type jpg, jpeg, png",
                        "type": "string",
                        "format": "binary"
                    }
                },
                "type": "object"
            },
...
            "PhotoDeletedResponse": {
                "title": "User Photo Delete response",
                "description": "User Photo Delete response",
                "properties": {
                    "message": {
                        "description": "User photo deleted",
                        "type": "string",
                        "example": "User photo deleted."
                    }
                },
                "type": "object"
            },
            "PhotoUpdatedResponse": {
                "title": "User Photo Updated response",
                "description": "User Photo Updated response",
                "properties": {
                    "message": {
                        "description": "User photo updated",
                        "type": "string",
                        "example": "User photo updated."
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}